Overhaul and simplify document model
Simplifies state tracking and validaitons in the model we use to track Form526 document submissions to Lighthouse
